requirements.txt
pandas==1.1.0
plotly==4.9.0
matplotlib==3.3.0
streamlit==0.73.1
Pillow==8.1.0
